backend/requirements.txt
absl-py==2.3.1
asgiref==3.11.0
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==5.2.8
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
image==1.5.33
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==2.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.2.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.5.0
mysqlclient==2.2.7
namex==0.1.0
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyJWT==2.10.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==3.2.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
